+++ src/helpers/messageAnalysis.ts
@@ -1,13 +1,8 @@
-/*
- * Message Analyzer
- *
- * This script contains useful functions for analyzing messages.
-*/
-
 import type { Message, User } from "discord.js";
 
 /**
  * Recognizes when a message is prepended with '> ', which tells the bot not to respond.
+ *
  * @param message the message to parse
  * @returns true if the message begins with '> ', false if otherwise
  */
@@ -17,6 +12,7 @@ export function isMarkedAsIgnore (message: Message): boolean {
 
 /**
  * Recognizes when a message is from the specified user.
+ *
  * @param message the message whose author to check
  * @param user the user to check if the message came from
  * @returns true if the message came from the user, false if otherwise
@@ -27,6 +23,7 @@ export function isFromUser (message: Message, user: User|null): boolean {
 
 /**
  * Recognizes when a message is empty (mostly likely an image or embed).
+ *
  * @param message the message to check
  * @returns true if the message has no text content, false if otherwise
  */
@@ -36,6 +33,7 @@ export function isEmpty (message: Message): boolean {
 
 /**
  * Recognizes when a message @ mentions the specified user.
+ *
  * @param message the message to check
  * @param user the user to see if the message mentions
  * @returns true if the message mentions the user, else if otherwise
